requirements.txt
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==3.0.2
numpy==1.16.1
Pillow==5.4.1
pkg-resources==0.0.0
pyparsing==2.3.1
python-dateutil==2.8.0
python-gflags==3.1.2
six==1.12.0
torch==1.0.1.post2
torchvision==0.2.1
